requirements.txt
alembic==1.7.7
amqp==5.1.1
async-timeout==4.0.2
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
django-environ==0.9.0
environ==1.0
Flask==2.1.0
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
password==0.2
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
redis==4.3.4
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.32
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.7.0
